fix: parity command in PROD
